test: Consolidate OS filtering tests to share expensive fixture
Group all tests using setup_hosts_for_operating_system_filtering into
TestOperatingSystemFiltering in test_filter_hosts.py so the fixture
(which creates 14 hosts across two accounts) runs only once instead
of three times across separate modules.
