+++ iqe-host-inventory-plugin/iqe_host_inventory/tests/rest/test_filter_tags.py
@@ -960,118 +960,9 @@ def test_filter_tags_by_system_profile_bootc_status_host_type(
     assert_tags_not_found(not_expected_tags, response.results)
 
 
-@pytest.mark.parametrize(
-    "params, expected_hosts",
-    [
-        (["[RHEL][version][lt][]=7.10"], [0]),
-        (["[RHEL][version][lte][]=7.10"], [0, 1]),
-        (["[RHEL][version][gte][]=7.10"], [1, 2]),
-        (["[RHEL][version][gt][]=7.10"], [2]),
-        (["[RHEL][version][lt][]=7.6"], [0]),
-        (["[RHEL][version][lte][]=7.6"], [0]),
-        (["[RHEL][version][gte][]=7.6"], [1, 2]),
-        (["[RHEL][version][gt][]=7.6"], [1, 2]),
-        (["[RHEL][version][lt][]=7"], []),
-        (["[RHEL][version][lte][]=7"], [0, 1]),
-        (["[RHEL][version][gte][]=7"], [0, 1, 2]),
-        (["[RHEL][version][gt][]=7"], [2]),
-        (["[RHEL][version][gt][]=8"], []),
-        (["[RHEL][version][gte][]=8"], [2]),
-        (["[RHEL][version][gte][]=7.10", "[RHEL][version][lte][]=7.10"], [1]),
-        (["[RHEL][version][gt][]=7", "[RHEL][version][lt][]=8"], []),
-        # TODO: Uncomment when https://issues.redhat.com/browse/RHINENG-18750 is fixed
-        # (["[RHEL][version][gte][]=7", "[RHEL][version][lte][]=8"], [0, 1, 2]),
-        (["[RHEL][version][eq][]=7"], [0, 1]),
-        (["[RHEL][version][eq][]=8"], [2]),
-        (["[RHEL][version][eq][]=7.10"], [1]),
-        (["[RHEL][version][eq][]=7.1"], []),
-        (["[RHEL][version][eq][]=7.5", "[RHEL][version][eq][]=8.0"], [0, 2]),
-        (["[RHEL][version][eq][]=7.6", "[RHEL][version][eq][]=8.0"], [2]),
-        (["[RHEL][version][eq][]=7", "[RHEL][version][eq][]=8.0"], [0, 1, 2]),
-        (["[RHEL][version][eq][]=7", "[RHEL][version][eq][]=8"], [0, 1, 2]),
-        (["[RHEL][version][eq][]=7", "[RHEL][version][eq][]=7"], [0, 1]),
-        (["[RHEL][version][eq][]=7.5", "[RHEL][version][gt][]=7.10"], [0, 2]),
-        (["[RHEL][version][eq][]=7.5", "[RHEL][version][gt][]=7.5"], [0, 1, 2]),
-        (
-            [
-                "[RHEL][version][eq][]=7.5",
-                "[RHEL][version][eq][]=7.10",
-                "[RHEL][version][eq][]=8.0",
-            ],
-            [0, 1, 2],
-        ),
-        (
-            ["[RHEL][version][eq][]=7", "[RHEL][version][eq][]=8", "[RHEL][version][eq][]=7.5"],
-            [0, 1, 2],
-        ),
-        (
-            ["[RHEL][version][eq][]=7", "[RHEL][version][eq][]=7.5", "[RHEL][version][eq][]=8.0"],
-            [0, 1, 2],
-        ),
-        (
-            ["[RHEL][version][eq][]=7.5", "[RHEL][version][gt][]=7.6", "[RHEL][version][lt][]=8"],
-            [0, 1],
-        ),
-        (["[CentOS Linux][version][lt][]=7.10"], [3]),
-        (["[CentOS Linux][version][lte][]=7.10"], [3, 4]),
-        (["[CentOS Linux][version][gte][]=7.10"], [4, 5]),
-        (["[CentOS Linux][version][gt][]=7.10"], [5]),
-        (["[CentOS Linux][version][gte][]=7.10", "[CentOS Linux][version][lte][]=7.10"], [4]),
-        (["[CentOS Linux][version][eq][]=7"], [3, 4]),
-        (["[CentOS Linux][version][eq][]=7.10"], [4]),
-        (["[CentOS Linux][version][eq][]=7.5", "[CentOS Linux][version][eq][]=8.0"], [3, 5]),
-        (["[RHEL][version][eq][]=7.5", "[CentOS Linux][version][eq][]=8.0"], [0, 5]),
-        (["=nil"], [6]),
-        (["=not_nil"], [0, 1, 2, 3, 4, 5]),
-        (["[]=nil", "[]=not_nil"], [0, 1, 2, 3, 4, 5, 6]),
-        (["[name][neq]=RHEL"], [3, 4, 5]),
-        (["[name][eq]=RHEL"], [0, 1, 2]),
-        (["[name][neq]=Centos Linux"], [0, 1, 2]),
-        (["[name][eq]=Centos Linux"], [3, 4, 5]),
-        (["[name][neq][]=CentOS Linux", "[name][neq][]=RHEL"], [0, 1, 2, 3, 4, 5]),
-        (["[name][eq][]=CentOS Linux", "[name][eq][]=RHEL"], [0, 1, 2, 3, 4, 5]),
-        (["[name][eq]=Centos Linux", "[CentOS Linux][version][eq][]=7.10"], [3, 4, 5]),
-        (["[name][eq]=Centos Linux", "[RHEL][version][eq][]=7.10"], [1, 3, 4, 5]),
-        (["[name][neq]=RHEL", "[RHEL][version][eq][]=7.10"], [1, 3, 4, 5]),
-        (["[name][eq]=RHEL", "[RHEL][version][eq][]=7.10"], [0, 1, 2]),
-    ],
-    ids=lambda param: "&".join(param) if param and isinstance(param[0], str) else None,
-)
-# TODO: Enable case_insensitive filtering  https://issues.redhat.com/browse/RHINENG-15761
-@pytest.mark.parametrize("case_insensitive", [False], ids=["case sensitive"])
-# @pytest.mark.parametrize(
-#     "case_insensitive", [False, True], ids=["case sensitive", "case insensitive"]
-# )
-def test_filter_tags_by_system_profile_operating_system(
-    setup_hosts_for_operating_system_filtering: tuple[list[HostOut], list[list[StructuredTag]]],
-    host_inventory: ApplicationHostInventory,
-    params: list[str],
-    expected_hosts: list[int],
-    case_insensitive: bool,
-):
-    """
-    https://issues.redhat.com/browse/RHINENG-10785
-
-    metadata:
-      requirements: inv-tags-get-list, inv-hosts-filter-by-sp-operating_system,
-      assignee: fstavela
-      importance: high
-      title: Inventory: filter tags by operating_system
-    """
-    _, tags = setup_hosts_for_operating_system_filtering
-    expected_tags = flatten(tags[i] for i in expected_hosts)
-    not_expected_tags = flatten(tags[i] for i in range(len(tags)) if i not in expected_hosts)
-
-    filter = [
-        f"[operating_system]{param.lower() if case_insensitive else param}" for param in params
-    ]
-    # There are too many "module" scoped hosts with tags, 50 isn't enough
-    response = host_inventory.apis.tags.get_tags_response(filter=filter, per_page=100)
-    log_response_tags_indices(tags, response)
-
-    assert response.count >= len(expected_tags)
-    assert_tags_found(expected_tags, response.results)
-    assert_tags_not_found(not_expected_tags, response.results)
+# NOTE: test_filter_tags_by_system_profile_operating_system has been moved to
+# test_filter_hosts.py::TestOperatingSystemFiltering to share the expensive
+# setup_hosts_for_operating_system_filtering fixture with other OS filtering tests.
 
 
 @pytest.mark.ephemeral
